+++ .gitignore
@@ -13,5 +13,5 @@ data_call/pycache
 data_call/__pycache__
 views/map.hbs
 data_call/table.json
-data_call/__pycache__/data.cpython-312.pyc
 routes/home.js
+data_call/__pycache__/data.cpython-312.pyc
