+++ app.js
@@ -1,6 +1,5 @@
 // --------------- code for connecting to the database ---------------
 
-
 require("dotenv").config();
 const bodyParser = require("body-parser");
 var bcrypt = require("bcryptjs");
@@ -53,8 +52,7 @@ var errorTable;
 async function fetchTableData() {
   // pull researcher table data from sql db, export it as json response
   var con = mysql.createConnection(sqlConfig);
-  var query1 =
-    "SELECT Devices.sn, Devices.pmHealth, Devices.sdHealth,Devices.onlne, Devices.dataFraction, CONCAT(ROUND(Devices.dataFraction*100,2),'%') AS dataFraction, Data.pm25, Data.pm10,  SUBSTRING(Data.timestamp,1,10) AS timestamp FROM Devices LEFT JOIN ( SELECT d1.* FROM Data d1 JOIN ( SELECT sn, MAX(timestamp) AS max_timestamp FROM Data GROUP BY sn ) d2 ON d1.sn = d2.sn AND d1.timestamp = d2.max_timestamp ) AS Data ON Data.sn = Devices.sn ORDER BY Devices.sn;";
+  var query1 = "SELECT Devices.sn, Devices.pmHealth, Devices.sdHealth,Devices.onlne, Devices.dataFraction, CONCAT(ROUND(Devices.dataFraction*100,2),'%') AS dataFraction, Data.pm25, Data.pm10,  SUBSTRING(Data.timestamp,1,10) AS timestamp FROM Devices LEFT JOIN ( SELECT d1.* FROM Data d1 JOIN ( SELECT sn, MAX(timestamp) AS max_timestamp FROM Data GROUP BY sn ) d2 ON d1.sn = d2.sn AND d1.timestamp = d2.max_timestamp ) AS Data ON Data.sn = Devices.sn ORDER BY Devices.sn;"
   await con
     .promise()
     .query(query1)
@@ -65,9 +63,8 @@ async function fetchTableData() {
       console.error(err);
     });
 
-  var query2 =
-    "SELECT Devices.sn, Devices.pmHealth, Devices.sdHealth, Devices.onlne, CONCAT(ROUND(Devices.dataFraction*100,2),'%') AS dataFraction FROM Devices WHERE Devices.sdHealth = 'ERROR' OR Devices.pmHealth='ERROR' OR Devices.onlne = 'offline' ORDER BY Devices.onlne, Devices.sdHealth DESC, Devices.pmHealth DESC;";
-
+  var query2 ="SELECT Devices.sn, Devices.pmHealth, Devices.sdHealth, Devices.onlne, CONCAT(ROUND(Devices.dataFraction*100,2),'%') AS dataFraction FROM Devices WHERE Devices.sdHealth = 'ERROR' OR Devices.pmHealth='ERROR' OR Devices.onlne = 'offline' ORDER BY Devices.onlne, Devices.sdHealth DESC, Devices.pmHealth DESC;";
+  
   await con
     .promise()
     .query(query2)
@@ -155,6 +152,12 @@ app.use("", homeRouter);
 const mapRouter = require("./routes/map.js");
 app.use("/map", mapRouter);
 
+app.get("/work-in-progress", (req, res) => {
+  res.render("work-in-progress", {
+    title: "Work in Progress",
+  });
+});
+
 // create route for the researcher table data
 app.get("/data", (req, res) => {
   res.json(tableData);
@@ -169,13 +172,15 @@ app.get("/researcher", async (req, res) => {
   res.json(addedResearchers);
 });
 
-//creates participant login page
-const loginRouter = require("./routes/login.js");
-app.use("/login", loginRouter);
+//creates data analysis testing page
+const dataAnalysisTestingRouter = require("./routes/data-analysis-testing.js");
+app.use("/data-analysis-testing", dataAnalysisTestingRouter);
 
 //creates success page
 const successRouter = require("./routes/success-page.js");
 app.use("/success-page", successRouter);
+
+//////////
 
 ///////////////////////////
 
@@ -430,77 +435,82 @@ app.use("/table", tableRouter);
 const lgRouter = require("./routes/logout.js");
 app.use("/logout", lgRouter);
 ////////////////////////////////////////////////////////////////
+//const router = express.Router();
+
+//route for the provisional page
+/*router.get("/work-in-progress", (req, res) => {
+  res.render("work-in-progress", {
+    title: "Work in Progress",
+  });
+});*/
+
+//Route for handling form submission for data analysis testing
+app.post("/data-analysis-testing", (req, res) => {
+  const monitorId = req.body.monitorId;
+
+  //For now, let's just render a page that displays a monitor ID prompt
+  res.render("data-analysis", { monitorId: monitorId });
+});
+///////////////////////////////////////////////////////
 
 ////////////////////////////////////////////////////////////////
 const router = express.Router();
 
-app.get("/monitorIds", async (req, res) => {
+//route for the provisional page
+/*router.get("/work-in-progress", (req, res) => {
+  res.render("work-in-progress", {
+    title: "Work in Progress",
+  });
+});*/
+
+//Route for handling form submission for data analysis testing
+//app.post("/data-analysis-testing", (req, res) => {
+//  const monitorId = req.body.monitorId;
+
+  //For now, let's just render a page that displays a monitor ID prompt
+ // res.render("data-analysis", { monitorId: monitorId });
+//});
+///////////////////////////////////////////////////////
+app.get('/monitorIds', async (req, res) => {
   try {
     const connection = await pool.promise().getConnection();
 
-    const [rows] = await connection.query("SELECT sn FROM Devices");
+    const [rows] = await connection.query('SELECT sn FROM Devices');
 
     connection.release();
 
-    const monitorIds = rows.map((row) => row.sn);
+    const monitorIds = rows.map(row => row.sn);
 
     res.json({ monitorIds });
   } catch (error) {
-    console.error("Error fetching monitor IDs from database:", error);
-    res.status(500).json({ error: "Internal server error" });
+    console.error('Error fetching monitor IDs from database:', error);
+    res.status(500).json({ error: 'Internal server error' });
   }
 });
-//////////////////////////////////////////////
+
 
 //////////////////////////////////////////////
+app.post("/data-analysis-testing", (req, res) => {
+  const monitorId = req.body.monitorId;
 
-const { PythonShell } = require("python-shell");
-let options = {
-  mode: "text",
-  pythonPath: ".venv/Scripts/python",
-  pythonOptions: ["-u"], // get print results in real-time
-};
+  console.log("Submitted monitor ID:", monitorId);
 
-// fix this
-async function fetchAQIData() {
-  const result = await new Promise((resolve, reject) => {
-    PythonShell.run("data_call/idontfuckingknow.py", options).then((result) => {
-      // if (err) {
-      //     console.error('Error fetching AQI data:', err);
-      //     return;
-      // }
-      console.log("hello");
-      aqiData = JSON.parse(result);
-      return aqiData;
-    });
-  });
-}
-fetchAQIData();
+  const isValidMonitorId = validateMonitorId(monitorId); 
 
-app.get("/aqiData", async (req, res) => {
-  let aqidata = await fetchAQIData();
-  console.log(aqidata);
-  res.json(aqidata);
+  if (isValidMonitorId) {
+      console.log("Valid monitor ID. Redirecting to success page...");
+      res.redirect("/success-page?monitorId=" + monitorId);
+  } else {
+      console.log("Invalid monitor ID. Redirecting back to data-analysis-testing page...");
+      res.redirect("/data-analysis-testing?error=invalid");
+  }
 });
 
-app.post("/changePMType", async (req, res) => {
-  const selectedPMType = req.body.pm_type;
-  console.log(selectedPMType);
-  let options = {
-    mode: "text",
-    pythonPath: ".venv/Scripts/python",
 
-    pythonOptions: ["-u"], // get print results in real-time
-    args: [selectedPMType],
-  };
 
-  await PythonShell.run("data_call/generateMap.py", options, (err, results) => {
-    if (err) throw err;
-    console.log("Map generation completed");
-  });
 
-  res.redirect("/map"); //redirects back to the map page
-});
+
+//////////////////////////////////////////////
 
 //Export the router
 module.exports = router;
