+++ data_call/data.py
@@ -406,22 +406,18 @@ from folium.plugins import HeatMap
 ###############################################################################################################
 
 
-
-
-def mapGeneration(data=None, pm_type='pm10'):
+def mapGeneration(data=None):
     if data is None:
         data = getAllRecent()
 
     # Generate a map with a central location of the Salton Sea area
     central_latitude = data['geo.lat'].mean()
     central_longitude = data['geo.lon'].mean()
-    m = folium.Map(location=[central_latitude, central_longitude], zoom_start=10,
-    tiles='https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}', attr='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, &copy; <a href="https://cartodb.com/attributions">CartoDB</a>')
-
+    m = folium.Map(location=[central_latitude, central_longitude], zoom_start=10,  tiles='https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}', attr = '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> '
+    'contributors, &copy; <a href="https://cartodb.com/attributions">CartoDB</a>')
 
     # Define air quality color ranges based on PM2.5 values
     color_ranges = {
-        'pm10': {
         #"Good": (0, 50),
         #"Moderate": (51, 100),
         #"Unhealthy for Sensitive Groups": (101, 150),
@@ -435,29 +431,18 @@ def mapGeneration(data=None, pm_type='pm10'):
         (151, 200): "red",
         (201, 300): "purple",
         (301, float('inf')): "maroon"
-        },
-        'pm25': {
-        (0, 50): "green",
-        (51, 100): "yellow",
-        (101, 150): "orange",
-        (151, 200): "red",
-        (201, 300): "purple",
-        (301, float('inf')): "maroon"
-        }
-     }
-
-    selected_color_range = color_ranges.get(pm_type)
+    }
 
     # Add markers for each monitor with appropriate air quality color
-    for index, row in data.dropna(subset=['geo.lat', 'geo.lon', pm_type]).iterrows():
+    for index, row in data.dropna(subset=['geo.lat', 'geo.lon', 'pm10']).iterrows():
         latitude = row['geo.lat']
         longitude = row['geo.lon']
-        pm_value = row[pm_type]
+        pm10_value = row['pm10']
 
         # Determine the air quality color based on the PM2.5 value
         marker_color = "blue"  # Default color if value doesn't fall into any range
-        for (min_value, max_value), color in selected_color_range.items():
-            if min_value <= pm_value <= max_value:
+        for (min_value, max_value), color in color_ranges.items():
+            if min_value <= pm10_value <= max_value:
                 marker_color = color
                 break
 
@@ -466,7 +451,8 @@ def mapGeneration(data=None, pm_type='pm10'):
         Serial Number: {row['sn']}<br>
         Latitude: {row['geo.lat']}<br>
         Longitude: {row['geo.lon']}<br>
-        {pm_type.upper()}: {pm_value}<br>
+        PM2.5: {row['pm25']}<br>
+        PM10: {row['pm10']}<br>
         Timestamp: {row['timestamp']}<br>
         """
 
@@ -481,29 +467,14 @@ def mapGeneration(data=None, pm_type='pm10'):
     fill_opacity=1,
 ).add_to(m)
 
-  # Adding Dropdown Menu
-    dropdown_html = f"""
-    <form action="/changePMType" method="post">
-    <select id="pm_dropdown" name = pm_type>
-        <option value="pm10" {'selected' if pm_type == 'pm10' else ''}>PM10</option>
-        <option value="pm25" {'selected' if pm_type == 'pm25' else ''}>PM2.5</option>
-    </select>
-    <button type="submit">Change</button>
-    </form>
-    """
 
-    m.get_root().html.add_child(folium.Element(dropdown_html))
-
-
-
-   # m.get_root().html.add_child(folium.Element(dropdown_html.format(pm10_selected=pm10_selected, pm25_selected=pm25_selected)))
     # Adding Legend
     legend_html = """
     <div style="position: fixed;
                 bottom: 50px; right: 50px; width: 230px; height: 155px;
                 border:3px solid black; z-index:9999; font-size:14px;
                 background-color:#f2f2f2;
-
+                /* Custom Ornate Border */
                 border-image: url('path/to/ornate-border.png') 30 round;
                 ">
     &nbsp;<b>Legend</b><br>
@@ -527,7 +498,6 @@ def mapGeneration(data=None, pm_type='pm10'):
 
     m.get_root().html.add_child(folium.Element(legend_html))
 
-
     # Save the map as an HTML file
     html_file_path = 'views/map.hbs'
     m.save(html_file_path)
@@ -535,14 +505,6 @@ def mapGeneration(data=None, pm_type='pm10'):
     # Open the HTML file in the default web browser
     webbrowser.open(html_file_path)
 
-    url = f'http://localhost:3000/map?pm_type={pm_type}'
-    print(url)
-  
-   # import sys
-   # if len(sys.argv) > 1:
-   #     mapGeneration(sys.argv[1])
-   # else:
-   #     mapGeneration()
 
 #####Added function to perform data analysis on the distribution of PM2.5 values#####
 
@@ -640,76 +602,98 @@ def calculate_aqi_for_all_monitors(pm25_data, pm10_data):
             aqi_pm25 = calculate_aqi(pm25_value)
             aqi_pm10 = calculate_aqi(pm10_value)
             aqi_values[monitor_id] = {'AQI_PM2.5': aqi_pm25, 'AQI_PM10': aqi_pm10}
-        # else:
-            # print(f"Missing data for monitor ID: {monitor_id}")
+        else:
+            print(f"Missing data for monitor ID: {monitor_id}")
     return aqi_values
+pm25_data = {
+    'MOD-PM-00637': 114.10195979899497, 'MOD-PM-00661': 36.25833333333333, 'MOD-PM-00673': 8.983333333333334,
+    'MOD-PM-00691': 19.633333333333333, 'MOD-PM-00690': 9.1125, 'MOD-PM-00645': 19.079166666666666,
+    'MOD-PM-00655': 130.61914572864322, 'MOD-PM-00692': 7.212500000000001, 'MOD-PM-00642': 11.845833333333333,
+    'MOD-PM-00666': 15.279166666666667, 'MOD-PM-00682': 32.483333333333334, 'MOD-PM-00687': 11.283333333333335,
+    'MOD-PM-00665': 40.733333333333334, 'MOD-PM-00678': 17.508333333333333, 'MOD-PM-00703': 5.370833333333334,
+    'MOD-PM-00676': 28.833333333333336, 'MOD-PM-00639': 2.816666666666667, 'MOD-PM-00695': 6.862500000000001,
+    'MOD-PM-00651': 20.062500000000004, 'MOD-PM-00677': float('nan'), 'MOD-PM-00704': 9.829166666666667,
+    'MOD-PM-00688': 1.3041666666666667, 'MOD-PM-00672': 5.7125, 'MOD-PM-00709': 6.075,
+    'MOD-PM-00656': 52.05781115879828, 'MOD-PM-00654': 4.9625, 'MOD-PM-00668': 4.2124999999999995,
+    'MOD-PM-00659': 37.28333333333334, 'MOD-PM-00674': 60.93459227467812, 'MOD-PM-00653': 31.27916666666667,
+    'MOD-PM-00641': 44.400000000000006, 'MOD-PM-00635': 9.316666666666668, 'MOD-PM-00683': 0.5333333333333334,
+    'MOD-PM-00711': 19.42916666666667, 'MOD-PM-00640': 14.175, 'MOD-PM-00675': 4.545833333333333,
+    'MOD-PM-00646': 51.29231759656653, 'MOD-PM-00696': 0.7583333333333334, 'MOD-PM-00652': 24.6125,
+    'MOD-PM-00660': 4.375000000000001
+}
 
+pm10_data = {
+    'MOD-PM-00637': 50.61443434343434, 'MOD-PM-00661': 25.336111111111112, 'MOD-PM-00673': 5.549074074074074,
+    'MOD-PM-00691': 4.423148148148148, 'MOD-PM-00690': 2.025, 'MOD-PM-00645': 15.852777777777776,
+    'MOD-PM-00655': 44.00833333333334, 'MOD-PM-00692': 3.713888888888889, 'MOD-PM-00642': 18.328703703703706,
+    'MOD-PM-00666': 5.110185185185185, 'MOD-PM-00682': 8.624074074074073, 'MOD-PM-00687': 13.337962962962962,
+    'MOD-PM-00665': 10.252777777777778, 'MOD-PM-00678': 5.8462962962962965, 'MOD-PM-00703': 1.9861111111111112,
+    'MOD-PM-00676': 32.931481481481484, 'MOD-PM-00639': 14.626851851851853, 'MOD-PM-00695': 3.003703703703704,
+    'MOD-PM-00651': 6.882407407407407, 'MOD-PM-00677': float('nan'), 'MOD-PM-00704': 2.55,
+    'MOD-PM-00688': 0.2898148148148148, 'MOD-PM-00672': 3.222222222222
+}
 
+aqi_values = calculate_aqi_for_all_monitors(pm25_data, pm10_data)
+for monitor_id, aqi_data in aqi_values.items():
+    print(f"Monitor ID: {monitor_id}")
+    if aqi_data['AQI_PM2.5'] is not None:
+        print(f"AQI_PM2.5: {aqi_data['AQI_PM2.5']:.2f}")
+    else:
+        print("AQI_PM2.5: N/A")
+    if aqi_data['AQI_PM10'] is not None:
+        print(f"AQI_PM10: {aqi_data['AQI_PM10']:.2f}")
+    else:
+        print("AQI_PM10: N/A")
+    print()
 def print_aqi_for_all_monitors(aqi_values):
     if aqi_values:
-        # for monitor_id, aqi_data in aqi_values.items():
-        #     print(f"Monitor ID: {monitor_id}")
-        #     if aqi_data['AQI_PM2.5'] is not None:
-        #         print(f"AQI_PM2.5: {aqi_data['AQI_PM2.5']:.2f}")
-        #     else:
-        #         print("AQI_PM2.5: N/A")
-        #     if aqi_data['AQI_PM10'] is not None:
-        #         print(f"AQI_PM10: {aqi_data['AQI_PM10']:.2f}")
-        #     else:
-        #         print("AQI_PM10: N/A")
-        #     print()
-        print(json.dumps(aqi_values))
+        for monitor_id, aqi_data in aqi_values.items():
+            print(f"Monitor ID: {monitor_id}")
+            if aqi_data['AQI_PM2.5'] is not None:
+                print(f"AQI_PM2.5: {aqi_data['AQI_PM2.5']:.2f}")
+            else:
+                print("AQI_PM2.5: N/A")
+            if aqi_data['AQI_PM10'] is not None:
+                print(f"AQI_PM10: {aqi_data['AQI_PM10']:.2f}")
+            else:
+                print("AQI_PM10: N/A")
+            print()
     else:
         print("No AQI data available for any monitors.")
 
+pm25_data = {
+    'MOD-PM-00637': 114.10195979899497, 'MOD-PM-00661': 36.25833333333333, 'MOD-PM-00673': 8.983333333333334,
+    'MOD-PM-00691': 19.633333333333333, 'MOD-PM-00690': 9.1125, 'MOD-PM-00645': 19.079166666666666,
+    'MOD-PM-00655': 130.61914572864322, 'MOD-PM-00692': 7.212500000000001, 'MOD-PM-00642': 11.845833333333333,
+    'MOD-PM-00666': 15.279166666666667, 'MOD-PM-00682': 32.483333333333334, 'MOD-PM-00687': 11.283333333333335,
+    'MOD-PM-00665': 40.733333333333334, 'MOD-PM-00678': 17.508333333333333, 'MOD-PM-00703': 5.370833333333334,
+    'MOD-PM-00676': 28.833333333333336, 'MOD-PM-00639': 2.816666666666667, 'MOD-PM-00695': 6.862500000000001,
+    'MOD-PM-00651': 20.062500000000004, 'MOD-PM-00677': float('nan'), 'MOD-PM-00704': 9.829166666666667,
+    'MOD-PM-00688': 1.3041666666666667, 'MOD-PM-00672': 5.7125, 'MOD-PM-00709': 6.075,
+    'MOD-PM-00656': 52.05781115879828, 'MOD-PM-00654': 4.9625, 'MOD-PM-00668': 4.2124999999999995,
+    'MOD-PM-00659': 37.28333333333334, 'MOD-PM-00674': 60.93459227467812, 'MOD-PM-00653': 31.27916666666667,
+    'MOD-PM-00641': 44.400000000000006, 'MOD-PM-00635': 9.316666666666668, 'MOD-PM-00683': 0.5333333333333334,
+    'MOD-PM-00711': 19.42916666666667, 'MOD-PM-00640': 14.175, 'MOD-PM-00675': 4.545833333333333,
+    'MOD-PM-00646': 51.29231759656653, 'MOD-PM-00696': 0.7583333333333334, 'MOD-PM-00652': 24.6125,
+    'MOD-PM-00660': 4.375000000000001
+}
+
+pm10_data = {
+    'MOD-PM-00637': 50.61443434343434, 'MOD-PM-00661': 25.336111111111112, 'MOD-PM-00673': 5.549074074074074,
+    'MOD-PM-00691': 4.423148148148148, 'MOD-PM-00690': 2.025, 'MOD-PM-00645': 15.852777777777776,
+    'MOD-PM-00655': 44.00833333333334, 'MOD-PM-00692': 3.713888888888889, 'MOD-PM-00642': 18.328703703703706,
+    'MOD-PM-00666': 5.110185185185185, 'MOD-PM-00682': 8.624074074074073, 'MOD-PM-00687': 13.337962962962962,
+    'MOD-PM-00665': 10.252777777777778, 'MOD-PM-00678': 5.8462962962962965, 'MOD-PM-00703': 1.9861111111111112,
+    'MOD-PM-00676': 32.931481481481484, 'MOD-PM-00639': 14.626851851851853, 'MOD-PM-00695': 3.003703703703704,
+    'MOD-PM-00651': 6.882407407407407, 'MOD-PM-00677': float('nan'), 'MOD-PM-00704': 2.55,
+    'MOD-PM-00688': 0.2898148148148148, 'MOD-PM-00672': 3.222222222222
+}
+
+aqi_values = calculate_aqi_for_all_monitors(pm25_data, pm10_data)
+print_aqi_for_all_monitors(aqi_values)
 
 
 
-from initializeDB import conn
-
-def fetch_data_from_database():
-    mydb = conn()
-    cursor = mydb.cursor()
-    cursor.execute("SELECT * FROM Data") 
-    columns = [col[0] for col in cursor.description]
-    data = cursor.fetchall()
-    df = pd.DataFrame(data, columns=columns)
-    return df
-
-def fetch_AQI_data_for_all_monitors():
-    mydb = conn()
-    cursor = mydb.cursor()
-
-    cursor.execute("SELECT * FROM Data")  
-    columns = [col[0] for col in cursor.description]
-    data = cursor.fetchall()
-    pm_data = pd.DataFrame(data, columns=columns)
-
-   
-   # def calculateAQI(pm25):
-        # stuff
-      #  return pm25  
-
-    pm_data['AQI'] = pm_data['pm25'].apply(calculate_aqi)
-
-    return pm_data
-
-pm_data = fetch_AQI_data_for_all_monitors()
-
-monitor_ids = pm_data['monitor_id'].unique()
-
-plt.figure(figsize=(12, 6))
-for monitor_id in monitor_ids:
-    data_subset = pm_data[pm_data['monitor_id'] == monitor_id]
-    plt.plot(data_subset['timestamp'], data_subset['AQI'], label=f"Monitor {monitor_id}")
-
-plt.title('AQI Values Over Time for Each Monitor ID')
-plt.xlabel('Timestamp')
-plt.ylabel('AQI')
-plt.xticks(rotation=45)
-plt.legend()
-plt.tight_layout()
-plt.show()
 
 
 def updateDataFractionForToday(serialNumber):
