--- data_call/idontfuckingknow.py
@@ -1,36 +1,0 @@
-import data as dc;
-import json
-
-
-pm25_data = {
-    'MOD-PM-00637': 114.10195979899497, 'MOD-PM-00661': 36.25833333333333, 'MOD-PM-00673': 8.983333333333334,
-    'MOD-PM-00691': 19.633333333333333, 'MOD-PM-00690': 9.1125, 'MOD-PM-00645': 19.079166666666666,
-    'MOD-PM-00655': 130.61914572864322, 'MOD-PM-00692': 7.212500000000001, 'MOD-PM-00642': 11.845833333333333,
-    'MOD-PM-00666': 15.279166666666667, 'MOD-PM-00682': 32.483333333333334, 'MOD-PM-00687': 11.283333333333335,
-    'MOD-PM-00665': 40.733333333333334, 'MOD-PM-00678': 17.508333333333333, 'MOD-PM-00703': 5.370833333333334,
-    'MOD-PM-00676': 28.833333333333336, 'MOD-PM-00639': 2.816666666666667, 'MOD-PM-00695': 6.862500000000001,
-    'MOD-PM-00651': 20.062500000000004, 'MOD-PM-00677': float('nan'), 'MOD-PM-00704': 9.829166666666667,
-    'MOD-PM-00688': 1.3041666666666667, 'MOD-PM-00672': 5.7125, 'MOD-PM-00709': 6.075,
-    'MOD-PM-00656': 52.05781115879828, 'MOD-PM-00654': 4.9625, 'MOD-PM-00668': 4.2124999999999995,
-    'MOD-PM-00659': 37.28333333333334, 'MOD-PM-00674': 60.93459227467812, 'MOD-PM-00653': 31.27916666666667,
-    'MOD-PM-00641': 44.400000000000006, 'MOD-PM-00635': 9.316666666666668, 'MOD-PM-00683': 0.5333333333333334,
-    'MOD-PM-00711': 19.42916666666667, 'MOD-PM-00640': 14.175, 'MOD-PM-00675': 4.545833333333333,
-    'MOD-PM-00646': 51.29231759656653, 'MOD-PM-00696': 0.7583333333333334, 'MOD-PM-00652': 24.6125,
-    'MOD-PM-00660': 4.375000000000001
-}
-
-pm10_data = {
-    'MOD-PM-00637': 50.61443434343434, 'MOD-PM-00661': 25.336111111111112, 'MOD-PM-00673': 5.549074074074074,
-    'MOD-PM-00691': 4.423148148148148, 'MOD-PM-00690': 2.025, 'MOD-PM-00645': 15.852777777777776,
-    'MOD-PM-00655': 44.00833333333334, 'MOD-PM-00692': 3.713888888888889, 'MOD-PM-00642': 18.328703703703706,
-    'MOD-PM-00666': 5.110185185185185, 'MOD-PM-00682': 8.624074074074073, 'MOD-PM-00687': 13.337962962962962,
-    'MOD-PM-00665': 10.252777777777778, 'MOD-PM-00678': 5.8462962962962965, 'MOD-PM-00703': 1.9861111111111112,
-    'MOD-PM-00676': 32.931481481481484, 'MOD-PM-00639': 14.626851851851853, 'MOD-PM-00695': 3.003703703703704,
-    'MOD-PM-00651': 6.882407407407407, 'MOD-PM-00677': float('nan'), 'MOD-PM-00704': 2.55,
-    'MOD-PM-00688': 0.2898148148148148, 'MOD-PM-00672': 3.222222222222
-}
-
-aqi_values = dc.calculate_aqi_for_all_monitors(pm25_data, pm10_data)
-
-dc.print_aqi_for_all_monitors(aqi_values)
-
