+++ routes/map.js
@@ -1,23 +1,14 @@
-const express = require("express");
-const router = express.Router();
-const { PythonShell } = require("python-shell");
+const express = require('express')
+const router = express.Router()
 
 // Getting all
-//router.get('/', (req,res) => {
-//    res.render('map', {
-//        title: 'AirPolice Map'
-//
-//    });
-//    res.status(200);
-//})
-
-router.get("/", (req, res) => {
-  res.render("map", {
-    title: "AirPolice Map",
-  });
-  res.status(200);
-});
+router.get('/', (req,res) => {
+    res.render('map', {
+        title: 'AirPolice Map'
+        
+    });
+    res.status(200); 
+})
 
 
-module.exports = router;
-
+module.exports = router
