+++ routes/success-page.js
@@ -1,16 +1,7 @@
-/*
 const express = require('express')
 const router = express.Router()
 
-//const { calculateAQI } = require('./');
-
-router.get('/', (req, res) => {
-    const monitorId = req.query.monitorId;
-
-    const pm25 = getPM25Value(monitorId); 
-    const pm10 = getPM10Value(monitorId); 
-    const aqi = calculateAQI(pm25, pm10);
-
+router.get('/', (req,res) => {
     
 
     //For now, let's just render a page that displays the success
@@ -19,35 +10,3 @@ router.get('/', (req, res) => {
 })
 
 module.exports = router;
-
-*/
-const express = require('express')
-const router = express.Router()
-
-
-router.get('/', (req,res) => {
-    
-    res.render("success-page", { title: 'SUCCESS PAGE ', monitorId : req.query.monitorId });
-    res.status(200);
-})
-
-//const { fetchData } = require('./data_call/data.py');
-
-router.get('/aqiData', (req, res) => {
-    const monitorId = req.query.monitorId;
-    const data = fetchData(monitorId);
-    res.json(data);
-});
-
-
-//const { fetchData } = require('./data');
-
-router.get('/aqiData', (req, res) => {
-    const data = fetchData();
-    res.json(data);
-});
-
-
-module.exports = router;
-
-
