+++ views/data-analysis-testing.hbs
@@ -1,0 +1,133 @@
+{{!-- BOOTSTRAP TEMPLATE FROM https://getbootstrap.com/docs/4.0/examples/checkout/ --}}
+
+<!doctype html>
+<html lang="en">
+  <head>
+    <link rel="icon" href="/docs/4.0/assets/img/favicons/favicon.ico">
+
+    <title>Participant Login</title>
+
+    <link rel="canonical" href="https://getbootstrap.com/docs/4.0/examples/checkout/">
+  </head>
+
+  <body class="bg-light">
+    <div class="container">
+      <div class="py-5 text-center">
+        <img class="d-block mx-auto mb-4" src="https://getbootstrap.com/docs/4.0/assets/brand/bootstrap-solid.svg" alt="" width="72" height="72">
+        <h2>Participant Login</h2>
+      </div>
+
+      <div class="row justify-content-center">
+        <div class="col-md-8 order-md-1">
+          <form class="needs-validation" novalidate>
+            <div class="mb-3">
+    <label for="monitorId">Please enter your Monitor ID</label>
+    <input type="text" class="form-control" id="monitorId" name="monitorId" placeholder="Monitor ID" required>
+    <div class="invalid-feedback">
+        Please enter a valid Monitor ID.
+    </div>
+</div>
+{{#if error}}
+<div class="alert alert-danger" role="alert">
+    {{error}}
+</div>
+{{/if}}
+
+
+            <hr class="mb-4">
+            <button class="btn btn-primary btn-lg btn-block" type="submit">Login</button>
+          </form>
+        </div>
+      </div>
+
+      <footer class="my-5 pt-5 text-muted text-center text-small">
+        <p class="mb-1">&copy; 2017-2018 Company Name</p>
+        <ul class="list-inline">
+          <li class="list-inline-item"><a href="#">Privacy</a></li>
+          <li class="list-inline-item"><a href="#">Terms</a></li>
+          <li class="list-inline-item"><a href="#">Support</a></li>
+        </ul>
+      </footer>
+    </div>
+
+
+
+
+
+
+
+
+
+    <!-- Bootstrap core JavaScript
+    ================================================== -->
+    <!-- Placed at the end of the document so the pages load faster -->
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+    <script>window.jQuery || document.write('<script src="../../assets/js/vendor/jquery-slim.min.js"><\/script>')</script>
+    <script src="../../assets/js/vendor/popper.min.js"></script>
+    <script src="../../dist/js/bootstrap.min.js"></script>
+    <script src="../../assets/js/vendor/holder.min.js"></script>
+    <script src="routes/data-analysis-testing.js"></script>
+    <script>
+      // Example starter JavaScript for disabling form submissions if there are invalid fields
+      (function() {
+        'use strict';
+
+        window.addEventListener('load', function() {
+          // Fetch all the forms we want to apply custom Bootstrap validation styles to
+          var forms = document.getElementsByClassName('needs-validation');
+
+          // Loop over them and prevent submission
+          var validation = Array.prototype.filter.call(forms, function(form) {
+            form.addEventListener('submit', function(event) {
+              if (form.checkValidity() === false) {
+                event.preventDefault();
+                event.stopPropagation();
+              }
+              form.classList.add('was-validated');
+            }, false);
+          });
+        }, false);
+      })();
+
+      function handleFormSubmission(event) {
+    event.preventDefault(); 
+    
+    const monitorIdInput = document.getElementById('monitorId');
+    const enteredMonitorId = monitorIdInput.value;
+
+    const validMonitorIds = [
+        "948", "1001", "976", "966", "993",
+        "1002", "977", "972", "998", "947",
+        "953", "1003", "956", "1000", "989",
+        "984", "1014", "987", "950", "978",
+        "982", "1006", "962", "1021", "346",
+        "988", "1015", "992", "999", "983",
+        "1020", "967", "965", "973", "979",
+        "970", "985", "995", "964", "952",
+        "949", "946", "994", "1022", "951",
+        "986", "957", "1007", "1023", "963"
+    ];
+
+    // Check if entered monitor ID is valid
+    const isValidMonitorId = validMonitorIds.some(id => enteredMonitorId.toUpperCase() === id.toUpperCase());
+
+    if (isValidMonitorId) {
+        alert("Correct Monitor ID.");
+    } else {
+        alert("Incorrect monitor ID. Please enter a valid Monitor ID.");
+    }
+}
+
+document.addEventListener('DOMContentLoaded', function() {
+    const form = document.querySelector('.needs-validation');
+    form.addEventListener('submit', handleFormSubmission);
+});
+{{#if error}}
+<div class="alert alert-danger" role="alert">
+    {{error}}
+</div>
+{{/if}}
+
+    </script>
+  </body>
+</html>
