+++ views/layouts/index.hbs
@@ -33,7 +33,7 @@
           </a>
 
           <ul class="dropdown-menu">
-            <li><a class="dropdown-item" href="/login">Participant</a></li>
+            <li><a class="dropdown-item" href="/data-analysis-testing">Participant</a></li>
             <li><a class="dropdown-item" href="/rlogin">Researcher</a></li>
             {{#if isPorterUser}}
                 <li><a class="dropdown-item" href="/invite">Invite</a></li>
