--- views/login.hbs
@@ -1,55 +1,0 @@
-{{!-- BOOTSTRAP TEMPLATE FROM https://getbootstrap.com/docs/4.0/examples/checkout/ --}}
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <link rel="icon" href="/docs/4.0/assets/img/favicons/favicon.ico">
-
-    <title>Participant Login</title>
-
-    <link rel="canonical" href="https://getbootstrap.com/docs/4.0/examples/checkout/">
-  </head>
-
-  <body class="bg-light">
-    <div class="container">
-      <div class="py-5 text-center">
-        <img class="d-block mx-auto mb-4" src="https://getbootstrap.com/docs/4.0/assets/brand/bootstrap-solid.svg" alt="" width="72" height="72">
-        <h2>Participant Login</h2>
-      </div>
-
-      <div class="row justify-content-center">
-        <div class="col-md-8 order-md-1">
-          <form class="needs-validation" novalidate>
-            <div class="mb-3">
-              <label for="monitorId">Please enter your Monitor ID</label>
-              <input type="text" class="form-control" id="monitorId" name="monitorId" placeholder="Monitor ID" required>
-              <div class="invalid-feedback">
-                  Please enter a valid Monitor ID.
-              </div>
-            </div>
-            <hr class="mb-4">
-            <button class="btn btn-primary btn-lg btn-block" type="submit">Login</button>
-          </form>
-        </div>
-      </div>
-
-      <footer class="my-5 pt-5 text-muted text-center text-small">
-        <p class="mb-1">&copy; 2017-2018 Company Name</p>
-        <ul class="list-inline">
-          <li class="list-inline-item"><a href="#">Privacy</a></li>
-          <li class="list-inline-item"><a href="#">Terms</a></li>
-          <li class="list-inline-item"><a href="#">Support</a></li>
-        </ul>
-      </footer>
-    </div>
-
-    
-
-    <script>
-      // monitorId is a variable stored in the route so here's how to access it in html:
-      // monitor ID: {{ monitorId }}
-    </script>
-
-
-  </body>
-</html>
